feat : "kafka를 통한 microservice 데이터 변경, order 발생 시 kafka connector를 통한 db추가"

+++ catalog-service/src/main/java/com/example/catalogservice/Infrastructure/consumer/KafkaConsumer.java
@@ -40,8 +40,8 @@ public class KafkaConsumer {
             );
         }
 
-        Long id = (Long)map.get("catalogId");
-        Integer quantity = (Integer)map.get("quantity");
+        Long id = ((Integer)map.get("catalogId")).longValue();
+        Integer quantity = (Integer) map.get("quantity");
 
         Catalog catalog = catalogRepository.findById(id)
                 .orElseThrow(() -> new ApplicationException(
